Added assets folder. Created dart files for home page, login page and account create
