+++ myapp/lib/src_exports.dart
@@ -1,1 +1,2 @@
 export 'src/gmaps.dart';
+export 'src/home.dart';
